64 bit compilation fix (Ubuntu 8.04) I did on the 10/06/08. Test results in 64bit are very similar to 32bit except off by one for various counters. (need to test if 32 bit results are affected)

+++ Network/IPv6/HdrExtFragProc.cc
@@ -103,8 +103,8 @@ bool HdrExtFragProc::assemblePacket(const int& nfrags, IPv6Datagram* frags,
 */
 //TODO This function is not correct
 IPv6Datagram** HdrExtFragProc::fragmentPacket(IPv6Datagram* pdu,
-                                              const size_t& path_mtu,
-                                              size_t& nfrags) const
+                                              const unsigned int& path_mtu,
+                                              unsigned int& nfrags) const
 {
   //Use the current frag_id
   assert(frag_id == 0); //Original packet was not a fragment itself?
